Move proto.Pcrs functionality to proto package
Move PCR related free functions in client into methods on Pcrs. This
makes the usage much more natural, and move non-client code out of the
client package.
