+++ client/import.go
@@ -39,7 +39,7 @@ func (k *Key) Import(blob *tpmpb.ImportBlob) ([]byte, error) {
 	}
 	defer tpm2.FlushContext(k.rw, handle)
 
-	unsealSession, err := newPCRSession(k.rw, PCRSelection(blob.Pcrs))
+	unsealSession, err := newPCRSession(k.rw, blob.Pcrs.PCRSelection())
 	if err != nil {
 		return nil, err
 	}
@@ -75,7 +75,7 @@ func (k *Key) ImportSigningKey(blob *tpmpb.ImportBlob) (key *Key, err error) {
 	if key.pubArea, _, _, err = tpm2.ReadPublic(k.rw, handle); err != nil {
 		return
 	}
-	if key.session, err = newPCRSession(k.rw, PCRSelection(blob.Pcrs)); err != nil {
+	if key.session, err = newPCRSession(k.rw, blob.Pcrs.PCRSelection()); err != nil {
 		return
 	}
 	return key, key.finish()
